[Index] Don’t verify mangled names of declarations in invalid decl contexts
The added test case crashed because we were trying to verify the mangled name of the `artifactID` parameter, because the parameter is not marked as invalid. But the mangled name of that parameter contains the name of the enum case, which is invalid. Thus, the mangled name contained errors and didn’t pass verification.

To fix the crash, don’t verify mangled names in decl contexts that are invalid.

rdar://128092323