[sending] Check for a nullptr in AssociatedTypeInference when calling computeFailureTypeWitness.
The problem happens when a type conforms to AsyncIterator.next when next in the
protocol returns its value as sending, but the witness does not have sending.
This results in the function subtyping rules rejecting the witness... but we
still leave in the witness as a nullptr... so we need to handle it here.

rdar://129300953
(cherry picked from commit 83bcf23dcafd4bc18145590c7e7de47ef60500e3)