[ScanDependencies] Make sure `canImport` resolution agrees with `import`
Fix the problem that when the only module can be found is an
invalid/out-of-date swift binary module, canImport and import statement
can have different view for if the module can be imported or not.

Now canImport will evaluate to false if the only module can be found for
name is an invalid swiftmodule, with a warning with the path to the
module so users will not be surprised by such behavior.

rdar://128876895
(cherry picked from commit 7d85aa423de39e4a8fa3b2dd5fea0a4a03361df0)