[Runtime] Add `swift_retainMultiple`/`swift_releaseMultiple`.
These are here to provide a way for `Array` to optimize reference counting
when copying; instead of having to call `swift_retain()` for every object
in the `Array`, it can just do `swift_retainMultiple()` for each chunk of
array storage.

The benefit of this is that we avoid a dyld stub if the retain calls are
inlined into user programs (since in that case we have to call into
`libswiftCore.dylib`).

rdar://126978101