Override default linker selection at build-time
The heuristic doesn't always work right and it's useful to be able to
build an internally consistent toolchain. This patch allows using the
`CLANG_DEFAULT_LINKER` setting to match the default behavior of the new
Swift driver and the old swift driver. Why CLANG_DEFAULT_LINKER you ask.
The new Swift driver falls back on clang to determine the default
linker. That means that the new Swift driver will use whichever linker
clang was configured to use. That means `CLANG_DEFAULT_LINKER`. I want
to ensure that both are tied to each other and one is not off doing
_other things_.

Why am I dropping the AmazonLinux2023 check? The heuristic mechanism
attempted to open `/etc/os-release` and check for the string `"Amazon
Linux 2023"`. This was fairly brittle and not exactly scalable. The
pretty name can also contain dot releases and build numbers, so this
check breaks. Instead, we should configure the Amazon Linux 2023
build to use lld by default and call it a day.