[DebugInfo] Dont generate info for __swift_async_resume_project_context or __swift_async_resume_get_context
These functions are alwaysinline and their instructions are created with line 0
as a location. As a result, the instructions all get inlined with line 0 into
funclets. In practice, this makes it so that the first instruction of the
funclet containing a debug location happens much later. This is important
because the inlined instructions do meaningful work, like overwriting the
contents of `fp - 8`.

If we try to set a breakpoint into this funclet, we *must* break before these
instructions, by allowing the line 0 locations to propagate, we fail to do.

This patch changes this by removing all debug locations from the helper
functions. By doing so, when they are inlined, instructions will inherit the
location of the call site.