[Concurrency] Calls that pass an isolated parameter matching the isolation of the context do not exit to nonisolated.
Setting the `mayExitToNonisolated` during isolated parameter checking was
done after a check that bails out early when the isolated argument matches
the isolated paramter of the caller, so the actor isolation checker was
accidentally considering the call as exiting to a nonisolated context. This
lead to bogus diagnostics during the region isolation pass.