Remove an unneeded function parameter by inlining its body.
There is only one call site of this helper function, so we are not gaining any
flexibility by doing this. It only makes it harder to reason about what the
function actually does without providing any benefit.

I am doing this before I make in the next commit an actual semantic changing
commit to ease review.