[Concurrency] Don't attempt to diagnose `GlobalConcurrency` issues with deserialized variable/properties
If the property comes from a different module the compiler shouldn't
attempt to diagnose `GlobalIsolation` problems based in the current
module flags otherwise it would create issues when i.e. a variable/property
from a module built with `-swift-version 5` that gets referenced
by a module that is built with `-swift-version 6` that has stricter
local concurrency requirements.

See https://forums.swift.org/t/swift-6-language-mode-being-passed-to-dependencies/72622
for further discussion.

A better solution would be to move the check to `ActorIsolationChecker`
or `DeclChecker` but that would be too risky for 6.0.

(cherry picked from commit aae15c33d40a4c18e9107910a2800fe42865193f)